backend/services/news-report-generator/requirements.txt
# Core Dependencies
feedparser==6.0.10
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0

# AI/ML
google-genai
# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Scheduler
apscheduler==3.10.4

# Validation
pydantic==2.5.0

# Utilities
python-dateutil==2.8.2
numpy==1.26.2
nltk==3.8.1
